+++ hardpy/pytest_hardpy/db/schema/v1.py
@@ -10,6 +10,7 @@ from typing import ClassVar
 from pydantic import BaseModel, ConfigDict, Field
 
 from hardpy.pytest_hardpy.utils import (
+    ChartType,
     ComparisonOperation as CompOp,
     Group,
     MeasurementType,
@@ -34,6 +35,7 @@ class CaseStateStore(IBaseResult):
     assertion_msg: str | None = None
     msg: dict | None = None
     measurements: list[NumericMeasurement | StringMeasurement] = []
+    chart: Chart | None = None
     attempt: int = 0
     group: Group
     dialog_box: dict = {}
@@ -45,6 +47,7 @@ class CaseRunStore(IBaseResult):
     assertion_msg: str | None = None
     msg: dict | None = None
     measurements: list[NumericMeasurement | StringMeasurement] = []
+    chart: Chart | None = None
     attempt: int = 0
     group: Group
     artifact: dict = {}
@@ -168,6 +171,20 @@ class StringMeasurement(IBaseMeasurement):
     casesensitive: bool = Field(default=True)
 
     comparison_value: str | None = Field(default=None)
+
+
+class Chart(BaseModel):
+    """Chart description."""
+
+    model_config = ConfigDict(extra="forbid")
+
+    type: ChartType = Field(default=ChartType.LINE)
+    title: str | None = Field(default=None)
+    x_label: str | None = Field(default=None)
+    y_label: str | None = Field(default=None)
+    marker_name: list[str | None] = Field(default=[])
+    x_data: list[list[int | float]] = Field(default_factory=lambda: [])  # noqa: PIE807
+    y_data: list[list[int | float]] = Field(default_factory=lambda: [])  # noqa: PIE807
 
 
 class OperatorData(BaseModel):